+++ jaxws-ri/runtime/rt/src/main/java-mr/9/com/sun/xml/ws/util/MrJarUtil.java
@@ -10,6 +10,9 @@
 
 package com.sun.xml.ws.util;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
 import java.security.AccessController;
 import java.security.PrivilegedAction;
 
@@ -36,5 +39,61 @@ public class MrJarUtil {
                 return noPool == null || Boolean.parseBoolean(noPool);
             }
         });
+    }
+
+    static InputStream getResourceAsStream(Class clazz, String resource) {
+        Package pkg = clazz.getPackage();
+        String fullpath = addPackagePath(resource, pkg);
+        InputStream is;
+
+        is = moduleResource(clazz, resource);
+        if (is != null) return is;
+
+        is = moduleResource(clazz, fullpath);
+        if (is != null) return is;
+
+        URL url = cpResource(clazz, resource);
+        if (url == null) url = cpResource(clazz, fullpath);
+
+        if (url == null) {
+            throw new UtilException("util.failed.to.find.handlerchain.file",
+                    clazz.getName(), resource);
+        }
+        try {
+            return url.openStream();
+        } catch (IOException e) {
+            throw new UtilException("util.failed.to.parse.handlerchain.file",
+                    clazz.getName(), resource);
+        }
+    }
+
+    private static URL cpResource(Class clazz, String name) {
+        URL url = clazz.getResource(name);
+        if (url == null) {
+            ClassLoader tccl = Thread.currentThread().getContextClassLoader();
+            url = tccl.getResource(name);
+        }
+        return url;
+    }
+
+    private static InputStream moduleResource(Class resolvingClass, String name) {
+        Module module = resolvingClass.getModule();
+        try {
+            InputStream stream = module.getResourceAsStream(name);
+            if (stream != null) {
+                return stream;
+            }
+        } catch(IOException e) {
+            throw new UtilException("util.failed.to.find.handlerchain.file",
+                    resolvingClass.getName(), name);
+        }
+        return null;
+    }
+
+    private static String addPackagePath(String file, Package pkg) {
+        String tmp = pkg.getName();
+        tmp = tmp.replace('.', '/');
+        tmp += "/" + file;
+        return tmp;
     }
 }
